# platform/scenario-b/apis/openapi/amm.yaml
openapi: 3.0.3
info:
  title: CBWeb3 Platform — AMM API (Scenario B)
  version: "2.0.0"
  description: |
    API Gateway endpoints for the Automated Market Maker (Scenario B).
    Covers cooperative liquidity provisioning, commit-reveal, swaps, quotes,
    and multi-pair PairRegistry management (005-cooperative-liquidity).
servers:
  - url: /api/v2/amm
    description: Scenario B AMM namespace
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
...
paths:
...
  /quote/exact-output:
    get:
      summary: Quote exact output swap
      operationId: getExactOutputQuote
      tags: [Quote]
      parameters:
        - name: pair
          in: query
          required: true
          schema:
            type: string
            example: "BRL-USD"
        - name: amount_out
          in: query
          required: true
          schema:
            type: string
            example: "1000000000000000000"
      responses:
        "200":
          description: Quote result
          content:
            application/json:
              schema:
                type: object
                properties:
                  required_input:
                    type: string
                  price_impact:
                    type: string
                  quote_timestamp:
                    type: integer
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /quote/cross-currency:
    get:
      summary: Generate cross-currency swap quote with 15s TTL
      operationId: getCrossCurrencyQuote
      tags: [Quote]
      parameters:
        - name: source_currency
          in: query
          required: true
          schema:
            type: string
            description: Native currency on source spoke
            example: "BRL"
        - name: target_currency
          in: query
          required: true
          schema:
            type: string
            description: Native currency on target spoke
            example: "ARS"
        - name: amount_out
          in: query
          required: true
          schema:
            type: string
            description: Exact amount beneficiary receives (wei)
            example: "1000000000000000000"
        - name: max_slippage_pct
          in: query
          required: false
          schema:
            type: number
            format: float
            description: Slippage tolerance (default 1%)
            example: 0.01
      responses:
        "200":
          description: Quote generated successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - quote_id
                  - pool_pair
                  - amount_out
                  - amount_in
                  - effective_rate
                  - fee_bps
                  - max_slippage_pct
                  - reserve_a_snapshot
                  - reserve_b_snapshot
                  - created_at
                  - valid_until
                  - time_remaining_seconds
                properties:
                  quote_id:
                    type: string
                    format: uuid
                    description: Unique quote identifier for validation
                    example: "550e8400-e29b-41d4-a716-446655440000"
                  pool_pair:
                    type: string
                    description: Pool used for swap (wrapped currencies)
                    example: "W-BRL-W-ARS"
                  amount_out:
                    type: string
                    description: Exact amount beneficiary receives (wei)
                    example: "1000000000000000000"
                  amount_in:
                    type: string
                    description: Required input amount including fee (wei)
                    example: "1003000000000000000"
                  effective_rate:
                    type: number
                    format: double
                    description: Exchange rate (amount_out / amount_in)
                    example: 0.997
                  fee_bps:
                    type: integer
                    description: Pool fee in basis points
                    example: 30
                  max_slippage_pct:
                    type: number
                    format: float
                    description: User tolerance for price movement
                    example: 0.01
                  reserve_a_snapshot:
                    type: string
                    description: Reserve A at quote time (audit trail, wei)
                    example: "10000000000000000000000"
                  reserve_b_snapshot:
                    type: string
                    description: Reserve B at quote time (audit trail, wei)
                    example: "10000000000000000000000"
                  created_at:
                    type: integer
                    format: int64
                    description: Quote generation timestamp (Unix epoch seconds)
                    example: 1716753600
                  valid_until:
                    type: integer
                    format: int64
                    description: Quote expiry timestamp (Unix epoch seconds)
                    example: 1716753615
                  time_remaining_seconds:
                    type: integer
                    description: Seconds until expiry (for client countdown)
                    example: 14
        "400":
          description: Bad request (missing required parameters)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Cannot generate quote (insufficient liquidity, pool not active)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Swap ─────────────────────────────────────────────────────────────
...
  /pool/{pair}/status:
    get:
      summary: Get pool status for a pair
      operationId: getPoolStatus
      tags: [Pool]
      parameters:
        - name: pair
          in: path
          required: true
          schema:
            type: string
            example: "BRL-USD"
      responses:
        "200":
          description: Pool status
          content:
            application/json:
              schema:
                type: object
                properties:
                  pair:
                    type: string
                  status:
                    type: string
                    enum: [ACTIVE, EMPTY, PENDING_COUNTERPART]
                  reserve_a:
                    type: string
                  reserve_b:
                    type: string

  # ── Liquidity Provision (commit-reveal) ───────────────────────────────
...
  /pairs/propose:
    post:
      summary: Propose a new currency pair (Central Bank of tokenA only)
      description: |
        The caller must be authenticated as the Central Bank whose on-chain address
        equals `IIdentityRegistry.getCentralBankOf(token_a_address)`.
        Creates the pair in PROPOSED status on-chain and in the DB.
        The Central Bank of tokenB must call `/pairs/confirm` to activate the pair.
      operationId: proposePair
      tags: [PairRegistry]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pair_id, token_a_address, token_b_address, amm_address, proposer_cb]
              properties:
                pair_id:
                  type: string
                  description: Human-readable pair identifier (must be unique)
                  example: "BRL-ARS"
                token_a_address:
                  type: string
                  description: tCeBM contract address for currency A (proposer's token)
                  example: "0x..."
                token_b_address:
                  type: string
                  description: tCeBM contract address for currency B
                  example: "0x..."
                amm_address:
                  type: string
                  description: Pre-deployed AutomatedMarketMaker contract for this pair
                  example: "0x..."
                proposer_cb:
                  type: string
                  description: Central Bank identifier of the proposer
                  example: "cb-brazil"
      responses:
        "201":
          description: Pair proposed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  pair_id:
                    type: string
                  status:
                    type: string
                    example: "PROPOSED"
                  tx_hash:
                    type: string
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller is not the Central Bank of token_a_address
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      code:
                        example: "NOT_CENTRAL_BANK_OF_TOKEN_A"
        "409":
          description: pair_id already exists
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      code:
                        example: "PAIR_ALREADY_EXISTS"
  /pairs/confirm:
    post:
      summary: Confirm a proposed pair (Central Bank of tokenB only)
      description: |
        The caller must be authenticated as the Central Bank whose on-chain address
        equals `IIdentityRegistry.getCentralBankOf(token_b_address)`.
        Transitions the pair from PROPOSED → ACTIVE on-chain and in the DB.
        The PairRouter cache is updated in real time via the PairRegistered event watcher (D10).
      operationId: confirmPair
      tags: [PairRegistry]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pair_id, confirmer_cb]
              properties:
                pair_id:
                  type: string
                  description: The pair identifier to confirm
                  example: "BRL-ARS"
                confirmer_cb:
                  type: string
                  description: Central Bank identifier of the confirmer
                  example: "cb-argentina"
      responses:
        "200":
          description: Pair confirmed and now ACTIVE
          content:
            application/json:
              schema:
                type: object
                properties:
                  pair_id:
                    type: string
                  status:
                    type: string
                    example: "ACTIVE"
                  tx_hash:
                    type: string
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller is not the Central Bank of token_b_address
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      code:
                        example: "NOT_CENTRAL_BANK_OF_TOKEN_B"
        "404":
          description: pair_id not found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      code:
                        example: "PAIR_NOT_FOUND"
        "409":
          description: Pair is already ACTIVE
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      code:
                        example: "PAIR_ALREADY_ACTIVE"